coding_challenge_project_v2/requirements.txt
Django==4.1.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
python-decouple==3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-filter==22.1